6-size.c

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
 */
 int main(void)
 {
-printf("Size of a char: %d bytes(s)\n", (int) sizeof(char));
-printf("Size of an int: %d bytes(s)\n", (int) sizeof(int));
-printf("Size of a long int: %d bytes(s)\n", (int) sizeof(long int));
-printf("Size of a long long int: %d bytes(s)\n", (int) sizeof(long long int));
-printf("Size of a float: %d bytes(s)\n", (int) sizeof(float));
+printf("Size of a char: %i bytes(s)\n", sizeof(char));
+printf("Size of an int: %i bytes(s)\n", sizeof(int));
+printf("Size of a long int: %i bytes(s)\n", sizeof(long int));
+printf("Size of a long long int: %i bytes(s)\n", sizeof(long long int));
+printf("Size of a float: %i bytes(s)\n", sizeof(float));
 return (0);
 }
